lab3/requirements.txt
fastapi==0.111.0
pydantic==2.7.1
python-dotenv==1.0.1
scikit-learn==1.4.2
uvicorn==0.29.0
uvloop==0.19.0
